requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
lxml==5.3.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
PyYAML==6.0.1
